Fix line numbers in generated code
Erbse was reporting wrong line numbers when an error occured in the
generated code. Turns out the cause is that it is necessary to tag each
newline with the :newline tag.

I add an explict match for all newlines, in order to introduce the text
and the :newline tag.

As a side effect, the output has a bit more newlines as well.